+++ app/endpoints/calendar.py
@@ -5,7 +5,7 @@ from fastapi import Depends, HTTPException
 from fastapi.responses import FileResponse
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.api import Module
+from app.utils.types.module import Module
 from app.core.config import Settings
 from app.cruds import cruds_calendar
 from app.dependencies import get_db, get_settings, is_user_a_member, is_user_a_member_of
