+++ app/endpoints/campaign.py
@@ -8,7 +8,7 @@ from fastapi.responses import FileResponse
 from pytz import timezone
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.api import Module
+from app.utils.types.module import Module
 from app.core.config import Settings
 from app.cruds import cruds_campaign, cruds_users
 from app.dependencies import get_db, get_request_id, get_settings, is_user_a_member_of
