+++ app/endpoints/groups.py
@@ -11,7 +11,7 @@ import uuid
 from fastapi import Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.api import Module
+from app.utils.types.module import Module
 from app.cruds import cruds_groups, cruds_users
 from app.dependencies import get_db, get_request_id, is_user_a_member_of
 from app.models import models_core
