+++ app/endpoints/loan.py
@@ -4,7 +4,7 @@ from typing import Sequence
 from fastapi import Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.api import Module
+from app.utils.types.module import Module
 from app.cruds import cruds_loan
 from app.dependencies import get_db, is_user_a_member, is_user_a_member_of
 from app.models import models_core, models_loan
