+++ app/endpoints/users.py
@@ -35,6 +35,7 @@ from app.utils.mail.mailworker import send_email
 from app.utils.tools import fuzzy_search_user, get_file_from_data, save_file_as_data
 from app.utils.types import standard_responses
 from app.utils.types.groups_type import AccountType, GroupType
+from app.utils.types.module import Module
 from app.utils.types.tags import Tags
 
 router = APIRouter()
